+++ inselect/gui/plugins/subsegment.py
@@ -72,8 +72,14 @@ class SubsegmentPlugin(Plugin):
         # Constrain rects to be within image
         rects = list(r.intersect(Rect(0.0, 0.0, 1.0, 1.0)) for r in rects)
 
-        # Replace the existing item
-        items[row:(1+row)] = [{'rect': r} for r in rects]
+        # Copy any existing metadata, rotation etc to the new items, update with
+        # new rects and replace the existing item
+        existing = items[row]
+        new_items = [None] * len(rects)
+        for index, rect in enumerate(image.to_normalised(rects)):
+            new_items[index] = existing.copy()
+            new_items[index]['rect'] = rect
+        items[row:(1+row)] = new_items
 
         # Segmentation image
         h, w = image.array.shape[:2]